css/css-anchor-position/anchor-center-005.html

<!DOCTYPE html>
<title>Tests auto-sizing of elements with `anchor-center`</title>
<link rel="help" href="https://drafts.csswg.org/css-anchor-position-1/#anchor-center">
<link rel="author" href="mailto:xiaochengh@chromium.org">
<link rel="stylesheet" href="/fonts/ahem.css">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>
<script src="support/test-common.js"></script>

<style>
.cb {
  position: relative;
  width: 150px;
  height: 150px;
  outline: 1px dashed gray;
}

.anchor {
  position: absolute;
  width: 50px;
  height: 50px;
  background: orange;
  anchor-name: --a;
}

.target {
  position: absolute;
  anchor-default: --a;
  font: 10px/1 Ahem;
  color: lime;
}

.align-x {
  justify-self: anchor-center;
  top: anchor(bottom);
}

.align-x::before{
  content: 'XX XX XX XX XX XX XX XX XX XX';
}

</style>

<body onload="checkLayoutForAnchorPos('.target')">
  <div class="cb">
    <div class="anchor"></div>
    <div class="target align-x" data-offset-x="0" data-expected-width="50"></div>
  </div>

  <div class="cb">
    <div class="anchor" style="left: 30px"></div>
    <div class="target align-x" data-offset-x="0" data-expected-width="110"></div>
  </div>

  <div class="cb">
    <div class="anchor" style="right: 30px"></div>
    <div class="target align-x" data-offset-x="40" data-expected-width="110"></div>
  </div>

  <div class="cb">
    <div class="anchor" style="right: 0"></div>
    <div class="target align-x" data-offset-x="100" data-expected-width="50"></div>
  </div>
</body>
